check_models.py
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()
api_key = os.getenv("GEMINI_API_KEY")

if not api_key:
    print("Error: GEMINI_API_KEY not found.")
else:
    genai.configure(api_key=api_key)
    try:
        print("Listing available PRO models:")
        count = 0
        for m in genai.list_models():
            if 'generateContent' in m.supported_generation_methods:
                if 'pro' in m.name.lower():
                     print(f"MODEL: {m.name}")
                     count += 1
        if count == 0:
            print("No models with 'pro' in the name found.")
            print("Listing ALL models:")
            for m in genai.list_models():
                 if 'generateContent' in m.supported_generation_methods:
                     print(f"MODEL: {m.name}")

    except Exception as e:
        print(f"Error listing models: {e}")
